♻️ [Refactor] : View별 파일분리 -추후 로직추가를 위해 각 View별로 파일을 분리했습니다. - 미션 선택할수 있는 뷰에 이미지 뷰를 추가했습니다.
